+++ src/components/content/SetupPage.jsx
@@ -1,0 +1,15 @@
+import React from "react";
+
+const SetupPage = () => {
+  return (
+    <div className='mainContent'>
+      <div className='mainInside bodyCopy'>
+        <div>Wika Network is built in the Polkadot ecosystem. You can reuse an existing Polkadot address if you already have one, or create a new one from scratch.</div>
+        <img className="mainThumb" src='dist/images/extension/thumb/white.svg' ></img>
+      </div>
+    </div>
+  )
+}
+
+export default SetupPage;
+
